+++ profile/exops/views.py
@@ -10,18 +10,18 @@ class ExportOpportunitiesBaseView(TemplateView):
     template_name_not_exops_user = 'exops/is-not-exops-user.html'
     template_name_error = 'exops/opportunities-retrieve-error.html'
 
-    opportunities = None
+    exops_data = None
     opportunities_retrieve_error = False
 
     def dispatch(self, request, *args, **kwargs):
         try:
-            self.opportunities = helpers.get_opportunities(request.user.id)
+            self.exops_data = helpers.get_exops_data(request.user.hashed_uuid)
         except HTTPError:
             self.opportunities_retrieve_error = True
         return super().dispatch(request, *args, **kwargs)
 
     def get_template_names(self, *args, **kwargs):
-        if self.opportunities is not None:
+        if self.exops_data is not None:
             template_name = self.template_name_exops_user
         elif self.opportunities_retrieve_error is True:
             template_name = self.template_name_error
@@ -33,7 +33,7 @@ class ExportOpportunitiesBaseView(TemplateView):
         search_url = settings.EXPORTING_OPPORTUNITIES_SEARCH_URL
         return {
             'exops_tab_classes': 'active',
-            'opportunities': self.opportunities,
+            'exops_data': self.exops_data,
             'EXPORTING_OPPORTUNITIES_SEARCH_URL': search_url,
         }